# src/docs/swagger.yaml
definitions:
...
  models.CandidateTime:
    properties:
      created_at:
        type: string
      end_time:
        type: string
      id:
        type: integer
      meeting_id:
        type: integer
      start_time:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
  models.CandidateTimeWithUserName:
    properties:
      end_time:
        type: string
      meeting_id:
        type: integer
      start_time:
        type: string
      user_name:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /api/restricted/candidate_times/available-time/{meeting_id}:
    get:
      consumes:
      - application/json
      description: Get an available time by meeting id
      parameters:
      - description: Meeting ID
        in: path
        name: meeting_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.CandidateTime'
        "400":
          description: Error message
          schema:
            type: string
      summary: Get an available time by meeting id
    put:
      consumes:
      - application/json
      description: Update an available time by meeting id
      parameters:
      - description: Meeting ID
        in: path
        name: meeting_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.CandidateTime'
        "400":
          description: Error message
          schema:
            type: string
      summary: Update an available time by meeting id
  /api/restricted/candidate_times/meeting/{meeting_id}:
    get:
      consumes:
      - application/json
      description: Get a CandidateTimeWithUserName by meeting id
      parameters:
      - description: Meeting ID
        in: path
        name: meeting_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.CandidateTimeWithUserName'
        "400":
          description: Error message
          schema:
            type: string
      summary: Get a CandidateTimeWithUserName by meeting id
  /api/restricted/candidate_times/new:
    post:
      consumes:
      - application/json
      description: Create a new candidate time
      parameters:
      - description: Details of candidate time
        in: body
        name: meeting
        required: true
        schema:
          $ref: '#/definitions/models.CandidateTime'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.CandidateTime'
        "400":
          description: Error message
          schema:
            type: string
      summary: Create a new candidate time
  /api/restricted/candidate_times/user/{user_id}/meeting/{meeting_id}:
    get:
      consumes:
      - application/json
      description: Get a candidate time by user id and meeting id
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: integer
      - description: Meeting ID
        in: path
        name: meeting_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.CandidateTime'
        "400":
          description: Error message
          schema:
            type: string
      summary: Get a candidate time by user id and meeting id
    put:
      consumes:
      - application/json
      description: Update a candidate time by user id and meeting id
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: integer
      - description: Meeting ID
        in: path
        name: meeting_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.CandidateTime'
        "400":
          description: Error message
          schema:
            type: string
      summary: Update a candidate time by user id and meeting id
...
swagger: "2.0"